Mark old rendering functionality as deprecated
